corrected some int that weren't doubles

+++ src/shapes/Square.java
@@ -3,19 +3,19 @@ import java.lang.Math;
 
 public class Square extends Rectangle {
 
-    protected int side;
+    protected double side;
 
-    public Square(int side) {
+    public Square(double side) {
         super(side, side);
         this.side = side;
     }
 
-    public int getArea(){
+    public double getArea(){
         System.out.println("This getArea is coming from square");
-        return (int) Math.pow(side, 2);
+        return Math.pow(side, 2);
     }
 
-    public int getPerimeter(){
+    public double getPerimeter(){
         System.out.println("\n\nThis getPerimeter is coming from square");
         return 4 * side;
     }
